Refactor sheet display utilities
