+++ commands/cmd_sheet.py
@@ -1,10 +1,10 @@
 from evennia import Command
 from django.db.utils import OperationalError
-from utils.sheet_display import (
+from utils.display import (
     display_pokemon_sheet,
     display_trainer_sheet,
-    get_status_effects,
 )
+from utils.display_helpers import get_status_effects
 from commands.command import get_max_hp
 from utils.xp_utils import get_display_xp
 from pokemon.stats import level_for_exp
